Update mac setup script

+++ setup-mac.sh
@@ -10,6 +10,7 @@ error() { echo -e "\033[1;31m[ERROR]\033[0m $1"; exit 1; }
 # ----------------------------------- SETUP -----------------------------------
 
 # --- homebrew setup
+log "# --- setting up package dependencies..."
 
 ### install Homebrew
 if ! command -v brew &>/dev/null; then
@@ -31,8 +32,10 @@ else
     error "brew.pkg file not found."
 fi
 
+export PKG_CONFIG_PATH="/opt/homebrew/opt/icu4c@76/lib/pkgconfig"
 
 # --- ZSH setup
+log "# --- setting up ZSH..."
 
 ### install OhMyZSH!
 if [[ -d "$HOME/.oh-my-zsh" ]]; then
@@ -61,11 +64,12 @@ install_zsh_plugin "https://github.com/romkatv/powerlevel10k.git" "$ZSH_CUSTOM/t
 install_zsh_plugin "https://github.com/zsh-users/zsh-autosuggestions" "$ZSH_CUSTOM/plugins/zsh-autosuggestions"
 
 # install nerd fonts
-log "Installing Nerd Fonts..."
-brew tap homebrew/cask-fonts
-brew install --cask font-mononoki-nerd-font
+# log "Installing Nerd Fonts..."
+# brew install --cask font-mononoki-nerd-font
 
 # --- install asdf
+log "# --- installing asdf..."
+
 if [[ ! -d "$HOME/.asdf" ]]; then
     log "Installing asdf..."
     git clone https://github.com/asdf-vm/asdf.git "$HOME/.asdf" --branch v0.14.1 || error "Failed to install asdf."
@@ -74,16 +78,21 @@ else
 fi
 
 # install SOPS
-curl -LO https://github.com/getsops/sops/releases/download/v3.9.1/sops-v3.9.1.darwin.arm64
-sudo mv sops-v3.9.1.darwin.arm64 /usr/local/bin/sops
-sudo chmod +x /usr/local/bin/sops
+# curl -LO https://github.com/getsops/sops/releases/download/v3.9.1/sops-v3.9.1.darwin.arm64
+# sudo mv sops-v3.9.1.darwin.arm64 /usr/local/bin/sops
+# sudo chmod +x /usr/local/bin/sops
 
 # ----------------------------------- CONFIGS -----------------------------------
 
 create_symlink() {
-    src=$1
-    dest=$2
-    ln -sf "$src" "$dest" && log "Created symlink: $src -> $dest"
+    local src=$1
+    local dest=$2
+
+    if [[ -L "$dest" && "$(readlink "$dest")" == "$src" ]]; then
+        log "Symlink already exists: $dest -> $src. Skipping."
+    else
+        ln -sf "$src" "$dest" && log "Created symlink: $src -> $dest"
+    fi
 }
 
 # --- setup symlinks to config files
@@ -92,26 +101,23 @@ if ! [[ -d ${HOME}/.config/helix/themes ]]; then
     mkdir ${HOME}/.config/helix/themes
 fi
 
-create_symlink $PWD/.bashrc ${HOME}
-create_symlink $PWD/.zshrc ${HOME}
-create_symlink $PWD/.gitconfig ${HOME}
-create_symlink $PWD/.p10k.zsh ${HOME}
-create_symlink $PWD/helix/config.toml ${HOME}/.config/helix/
-create_symlink $PWD/helix/themes/night_owl.toml ${HOME}/.config/helix/themes/
-create_symlink $PWD/tilda/config_0 ${HOME}/.config/tilda/
+create_symlink $PWD/.bashrc ${HOME}/.bashrc
+create_symlink $PWD/.zshrc ${HOME}/.zshrc
+create_symlink $PWD/.gitconfig ${HOME}/.gitconfig
+create_symlink $PWD/.p10k.zsh ${HOME}/.p10k.zsh
+create_symlink $PWD/helix/config.toml ${HOME}/.config/helix/config.toml
+create_symlink $PWD/helix/languages.toml ${HOME}/.config/helix/languages.toml
+create_symlink $PWD/helix/themes/night_owl.toml ${HOME}/.config/helix/themes/night_owl.toml
 
 # ----------------------------------- INSTALLATION -----------------------------------
 
-# --- install asdf plugins and tools
 
-declare -A tools=(
-    [python]=3.12.8
-    [terraform]=1.10.2
-    [gcloud]=503.0.0
-    [nodejs]=v22.12.0
-    [rust]=1.83.0
-    [postgres]=17.2
-)
+# --- install asdf plugins and tools
+log "# --- installing asdf plugins and tools..."
+
+# Arrays for tools and versions
+tools=("python" "terraform" "gcloud" "nodejs" "rust" "postgres")
+versions=("3.12.8" "1.10.2" "503.0.0" "v22.12.0" "1.83.0" "17.2")
 
 # Function to add and install asdf plugins/tools
 install_tool() {
@@ -120,42 +126,79 @@ install_tool() {
 
     # Check if plugin exists
     if asdf plugin-list | grep -q "^${tool}$"; then
-        log "Plugin '${tool}' already exists. Skipping plugin addition."
+        echo "[LOG]: Plugin '${tool}' already exists. Skipping plugin addition."
     else
-        log "Adding plugin '${tool}'..."
-        asdf plugin-add "$tool" || error "Failed to add plugin '${tool}'."
+        echo "[LOG]: Adding plugin '${tool}'..."
+        asdf plugin-add "$tool" || {
+            echo "[ERROR]: Failed to add plugin '${tool}'." >&2
+            exit 1
+        }
     fi
 
     # Install tool version
     if asdf list "$tool" | grep -q "$version"; then
-        log "Tool '${tool}' version '${version}' is already installed."
+        echo "[LOG]: Tool '${tool}' version '${version}' is already installed."
     else
-        log "Installing '${tool}' version '${version}'..."
-        asdf install "$tool" "$version" || error "Failed to install '${tool}' version '${version}'."
+        echo "[LOG]: Installing '${tool}' version '${version}'..."
+        asdf install "$tool" "$version" || {
+            echo "[ERROR]: Failed to install '${tool}' version '${version}'." >&2
+            exit 1
+        }
     fi
 
     # Set tool version globally
-    log "Setting '${tool}' version '${version}' globally."
-    asdf global "$tool" "$version" || error "Failed to set '${tool}' version globally."
+    echo "[LOG]: Setting '${tool}' version '${version}' globally."
+    asdf global "$tool" "$version" || {
+        echo "[ERROR]: Failed to set '${tool}' version globally." >&2
+        exit 1
+    }
 }
 
 # Install all tools
-for tool in "${!tools[@]}"; do
-    install_tool "$tool" "${tools[$tool]}"
+for i in "${!tools[@]}"; do
+    install_tool "${tools[$i]}" "${versions[$i]}"
 done
 
-log "All tools have been successfully installed and configured."
+echo "[LOG]: All tools have been successfully installed and configured."
+
+
+# --- install Helix
+log "# --- installing helix..."
+
+# Check if Helix is already installed
+if [[ -d "$HOME/src/helix" ]]; then
+    log "Helix is already installed. Skipping installation."
+else
+    log "Installing Helix.."
+
+    # Create the source directory if it doesn't exist
+    if [[ ! -d "$HOME/src" ]]; then
+        log "Creating source directory: $HOME/src"
+        mkdir -p "$HOME/src" || error "Failed to create source directory."
+    fi
+
+    # Clone the Helix repository
+    log "Cloning Helix repository into $HOME/src/helix..."
+    git clone git@github.com:helix-editor/helix.git "$HOME/src/helix" || error "Failed to clone Helix repository."
+
+    # Install Helix using Cargo
+    log "Installing Helix using Cargo..."
+    cargo install --path "$HOME/src/helix/helix-term" --locked || error "Failed to install Helix."
+
+    log "Helix installation completed successfully."
+fi
 
 # --- install helix language servers
+log "# --- installing helix language servers..."
 
-### Python LSP installation ---
+### Ruff installation
 if command -v pip >/dev/null; then
-    if pip show python-lsp-server >/dev/null 2>&1; then
-        log "Python language server (pylsp) is already installed. Skipping."
+    if pip show ruff >/dev/null 2>&1; then
+        log "Ruff (ruff) is already installed. Skipping."
     else
-        log "Installing Python language server (pylsp)..."
-        pip install "python-lsp-server[all]" || error "Failed to install Python language server."
-        log "Python language server installed successfully."
+        log "Installing Ruff (ruff)..."
+        pip install ruff || error "Failed to install Ruff."
+        log "Ruff installed successfully."
     fi
 else
     error "pip is not installed. Please install Python first."
@@ -175,6 +218,7 @@ else
 fi
 
 # --- install dev utils
+log "# --- installing dev utils..."
 
 # Ensure Cargo is available
 if ! command -v cargo >/dev/null; then
